Fix video (VoD) download from video player
The download directory is inside Android/data so the app does not ask for storage permission.
This removes the check for the permission (removed with scoped storage) and prevents
the app from throwing missing permission error.

Fixes an issue with the result of binarysearch if the exact match is not found that crashes the app.

(cherry picked from commit 1d5e805bb31f539ae341f45ce7aac691c0dfaaec)